[medium] Don't mutate existing array item
This is a common gotcha in JS: arrays and objects are passed by
reference (vs. pass by value). What this means is that when you
copy certain elements, you have to watch out for whether you're
creating a clone or actually referring to the same old element.

In this case, the `newShoppingCart` is not the same array as
`shoppingCart`, but each item in it (since they are objects) are
not duplicated, and refer to the same underlying object.

A simple case you can use to see this in action is like this:

```
const array1 = [{colour: "red"}, {colour: "blue"}, {colour: "green"}];
// btw, you can quickly copy arrays like:
// const array2 = [...array1];
const array2 = [];
for (let i = 0; i < array1.length; i++) {
  array2.push(array1[i]);
}

array2[0].colour = "yellow";

console.log(array1);
console.log(array2);
```

You should see that the output of the two logs are the same, because
array1 and array2 are the same collection of object *references*
(like an array of pointers in other languages).

--------

The code change in this commit is one way you can do it safely, and
another approach which is a bit longer could be like this:

```
const oldItem = newShoppingCart[itemIndex];
const newShoppingCart = [
   ...shoppingCart.slice(0, itemIndex),
   {itemId: oldItem.itemId, quantity: oldItem.quantity + 1},
   ...shoppingCart.slice(itemIndex + 1)
]
```