[small] setIsFetching inside async function for ProductDetail
Same as the other one, you don't have a *lot* of guarantees about how
or when your async function runs, so one of the ways to guarantee your
state changes happen at the right time is to put them in the
async function.

i.e. if you leave the setIsFetching(false) after the fetchData() call,
there's a chance that it will execute before the fetchData() function
even *starts*, which is definitely not what you want.