[medium] Use allProducts.filter to filter category/search
You can make your logic (and your life) simpler by using the
Array.filter function --> this way you don't have to keep making new
arrays and populating them with the items you want.

filter is a convenient way to do this that lets you write one
function to decide whether you want to keep the item in the array
or not.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/filter